mailer/requirements.txt
requests
python-firebase
Mako
